Room is joined inside of meeting index page
